requirements_mllm.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.3.3
certifi==2025.1.31
distro==1.9.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-core==2.4.1
google-cloud-translate==3.15.0
googleapis-common-protos==1.63.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
load-dotenv==0.1.0
numpy==2.2.2
openai==1.62.0
opencv-python==4.11.0.86
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
rsa==4.9
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
